+++ clang/docs/ReleaseNotes.rst
@@ -278,6 +278,10 @@ Python Binding Changes
 
 OpenMP Support
 --------------
+- Added support for 'omp assume' directive.
+- Added support for 'omp scope' directive.
+- Added support for allocator-modifier in 'allocate' clause.
+- Added support for 'omp dispatch' directive.
 
 Improvements
 ^^^^^^^^^^^^